Rename ConfigurableGremlinQuerySourceExtensions class in Providers.Neptune.
